+++ pymongo/client_options.py
@@ -53,7 +53,7 @@ def _parse_read_preference(options):
 def _parse_write_concern(options):
     """Parse write concern options."""
     concern = options.get('w')
-    wtimeout = options.get('wtimeout')
+    wtimeout = options.get('wtimeout', options.get('wtimeoutms'))
     j = options.get('j', options.get('journal'))
     fsync = options.get('fsync')
     return WriteConcern(concern, wtimeout, j, fsync)